fix: 프로젝트 구조 변경, index.jsp URL 변경
