+++ src/main/java/com/brokurly/dto/PointLogEarningDto.java
@@ -1,4 +1,4 @@
-package com.brokurly.entity;
+package com.brokurly.dto;
 
 import lombok.*;
