+++ src/main/java/com/brokurly/dto/PointLogExpDto.java
@@ -1,4 +1,4 @@
-package com.brokurly.entity;
+package com.brokurly.dto;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -8,9 +8,9 @@ import lombok.ToString;
 import java.util.Date;
 
 @Getter @Setter @ToString @EqualsAndHashCode
-public class Point {
-    private int pointNo;
-    private String custId;
+public class PointLogExpDto {
     private int pointAmt;
-    private Date expirDt;
+    private Date procDt;
+    private String pointSpec;
+    private String pointStat;
 }
