+++ src/main/java/com/brokurly/repository/PointDao.java
@@ -1,12 +1,7 @@
 package com.brokurly.repository;
 
-import com.brokurly.entity.Point;
-import com.brokurly.entity.PointLogExpDto;
-import com.brokurly.entity.PointLogUsageDto;
+import com.brokurly.domain.Point;
 import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Param;
-
-import java.util.List;
 
 @Mapper
 public interface PointDao {
